+++ packages/google_ml_kit/example/lib/vision_detector_views/label_detector_view.dart
@@ -1,5 +1,10 @@
+import 'dart:io' as io;
+
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:google_mlkit_image_labeling/google_mlkit_image_labeling.dart';
+import 'package:path/path.dart';
+import 'package:path_provider/path_provider.dart';
 
 import 'camera_view.dart';
 import 'painters/label_detector_painter.dart';
@@ -17,6 +22,13 @@ class _ImageLabelViewState extends State<ImageLabelView> {
   String? _text;
 
   @override
+  void initState() {
+    super.initState();
+
+    _initializeLabeler();
+  }
+
+  @override
   void dispose() {
     _canProcess = false;
     _imageLabeler.close();
@@ -29,26 +41,31 @@ class _ImageLabelViewState extends State<ImageLabelView> {
       title: 'Image Labeler',
       customPaint: _customPaint,
       text: _text,
-      onImage: (inputImage) {
-        // comment this line if you want to use custom model
-        processImageWithDefaultModel(inputImage);
-        // uncomment this line if you want to use custom model
-        // processImageWithRemoteModel(inputImage);
-      },
+      onImage: processImage,
     );
   }
 
-  Future<void> processImageWithDefaultModel(InputImage inputImage) async {
-    _imageLabeler = ImageLabeler(options: ImageLabelerOptions());
-    processImage(inputImage);
-  }
+  void _initializeLabeler() async {
+    // uncomment next line if you want to use the default model
+    // _imageLabeler = ImageLabeler(options: ImageLabelerOptions());
 
-  // Add the tflite model in android/src/main/assets
-  Future<void> processImageWithRemoteModel(InputImage inputImage) async {
-    final options = FirebaseLabelerOption(
-        confidenceThreshold: 0.5, modelName: 'bird-classifier');
-    _imageLabeler = ImageLabeler(options: options);
-    processImage(inputImage);
+    // uncomment next lines if you want to use a local model
+    // make sure to add tflite model to assets/ml
+    // final path = 'assets/ml/lite-model_aiy_vision_classifier_birds_V1_3.tflite';
+    final path = 'assets/ml/object_labeler.tflite';
+    final customModelPath = await _getModel(path);
+    _imageLabeler = ImageLabeler(
+        options: LocalLabelerOptions(customModelPath: customModelPath));
+
+    // uncomment next lines if you want to use a remote model
+    // make sure to add model to firebase
+    // final modelName = 'bird-classifier';
+    // final response =
+    //     await FirebaseImageLabelerModelManager().downloadModel(modelName);
+    // print('Downloaded: $response');
+    // final options =
+    //     FirebaseLabelerOption(confidenceThreshold: 0.5, modelName: modelName);
+    // _imageLabeler = ImageLabeler(options: options);
   }
 
   Future<void> processImage(InputImage inputImage) async {
@@ -76,5 +93,20 @@ class _ImageLabelViewState extends State<ImageLabelView> {
     if (mounted) {
       setState(() {});
     }
+  }
+
+  Future<String> _getModel(String assetPath) async {
+    if (io.Platform.isAndroid) {
+      return 'flutter_assets/$assetPath';
+    }
+    final path = '${(await getApplicationSupportDirectory()).path}/$assetPath';
+    await io.Directory(dirname(path)).create(recursive: true);
+    final file = io.File(path);
+    if (!await file.exists()) {
+      final byteData = await rootBundle.load(assetPath);
+      await file.writeAsBytes(byteData.buffer
+          .asUint8List(byteData.offsetInBytes, byteData.lengthInBytes));
+    }
+    return file.path;
   }
 }